==> Notebooks/05_feature_selection.ipynb <==
# Imports

import pandas as pd
import numpy as np
from numpy import where
import matplotlib.pyplot as plt

import sklearn as sk
from sklearn.linear_model import LogisticRegression
from sklearn import svm
from sklearn.ensemble import RandomForestClassifier
from sklearn.neural_network import MLPClassifier
from sklearn.model_selection import train_test_split

from sklearn import metrics
import seaborn as sns
from sklearn.inspection import permutation_importance
from sklearn.ensemble import GradientBoostingClassifier
from pycm import *
from sklearn.experimental import enable_iterative_imputer
from sklearn.impute import IterativeImputer
from sklearn import tree
from sklearn.model_selection import RepeatedStratifiedKFold
from sklearn.model_selection import cross_val_score
from sklearn.neighbors import KNeighborsClassifier
from sklearn.naive_bayes import GaussianNB

from collections import Counter
import xgboost as xgb
from xgboost import XGBClassifier
from lightgbm import LGBMClassifier
from catboost import CatBoostClassifier

from boruta import BorutaPy
import ppscore as pps

# Feature Selection

X_test = pd.read_excel(f'../processed_data/outliers_removed/imputed/knn/x_test.xlsx')
X_train = pd.read_excel(f'../processed_data/outliers_removed/imputed/knn/x_train.xlsx')
y_train = pd.read_excel(f'../processed_data/outliers_removed/imputed/knn/y_train.xlsx').values.ravel()
y_test = pd.read_excel(f'../processed_data/outliers_removed/imputed/knn/y_test.xlsx').values.ravel()

X_test.columns

# Barouta

# define random forest classifier, with utilising all cores and
# sampling in proportion to y labels


def run_barouta(model, X_test, y_test):
    # define Boruta feature selection method
    feat_selector = BorutaPy(model, n_estimators='auto', verbose=2, random_state=42)

    # find all relevant features - 5 features should be selected
    feat_selector.fit(X_test.values, y_test)

    # call transform() on X to filter it down to selected features
    X_filtered = feat_selector.transform(X_test.values)

    # zip my names, ranks, and decisions in a single iterable
    feature_ranks = list(zip(X_test.columns, 
                             feat_selector.ranking_, 
                             feat_selector.support_))

    display(feature_ranks)
    feature_ranks_dict = {}

    for i in feature_ranks:
        name, rank, valid = i
        feature_ranks_dict[name] = rank

    return feature_ranks_dict

rf = RandomForestClassifier(n_jobs=-1, class_weight='balanced', max_depth=5)
barouta_rf_dict = run_barouta(rf, X_test, y_test)
barouta_rf_dict

# Sci Kit Learn

rf = RandomForestClassifier(n_jobs=-1, class_weight='balanced', max_depth=5)
rf.fit(X_train, y_train)

scikitlearn_features = pd.DataFrame({'feature_importance': rf.feature_importances_}, index=X_train.columns)
scikitlearn_features['SciKitLearn: RandomForest'] = scikitlearn_features['feature_importance'].rank(method='dense', ascending=False)

scikitlearn_features = scikitlearn_features.drop('feature_importance', axis=1)
scikitlearn_features

# Predictive Power Score

df = X_train.copy()
df['mean_happiness'] = y_train

pps_df = pps.predictors(df, "mean_happiness")
pps_df['pps_rank'] = pps_df['ppscore'].rank(method='dense', ascending=False)

# Correlation

corr_dict = {}

for col in X_train.columns:
    corr = np.corrcoef(X_train[col], y_train)[0][1]
    corr_dict[col] = corr

corr_dict = {key: rank for rank, key in enumerate(sorted(corr_dict, key=corr_dict.get), 1)}
corr_dict

# Comparison

feature_comparison_df = pps_df[['x', 'pps_rank']].set_index('x')
feature_comparison_df['Barouta: RandomForest'] = feature_comparison_df.index.map(barouta_rf_dict)

feature_comparison_df['correlation'] = feature_comparison_df.index.map(corr_dict)

feature_comparison_df['average_rank'] = feature_comparison_df.mean(axis=1)
feature_comparison_df = feature_comparison_df.sort_values('average_rank')

feature_comparison_df.head(35)

textfile = open("features.txt", "w")
for element in feature_comparison_df.index:
    textfile.write(element + "\n")
textfile.close()
