==> Notebooks/06_no_of_features.ipynb <==
import pandas as pd
import numpy as np
from numpy import where
import matplotlib.pyplot as plt

import sklearn as sk
from sklearn.linear_model import LogisticRegression
from sklearn import svm
from sklearn.ensemble import RandomForestClassifier
from sklearn.neural_network import MLPClassifier
from sklearn.model_selection import train_test_split

from sklearn import metrics
import seaborn as sns
from sklearn.inspection import permutation_importance
from sklearn.ensemble import GradientBoostingClassifier
from sklearn.experimental import enable_iterative_imputer
from sklearn.impute import IterativeImputer
from sklearn import tree
from sklearn.model_selection import RepeatedStratifiedKFold
from sklearn.model_selection import cross_val_score
from sklearn.neighbors import KNeighborsClassifier
from sklearn.naive_bayes import GaussianNB

from collections import Counter
import xgboost as xgb
from xgboost import XGBClassifier
from lightgbm import LGBMClassifier
from catboost import CatBoostClassifier
from sklearn.impute import SimpleImputer
from sklearn.impute import KNNImputer

from functions import make_models, compare_models

# Importing features

with open("features.txt") as f:
    features = f.readlines()
    features = [f.replace('\n', '') for f in features]

feat_added = []
final_feats = []

for feat in features:
    og_feat = feat.replace('_no_outliers', '')
    if og_feat in feat_added:
        pass
    else:
        final_feats.append(feat)
        feat_added.append(og_feat)
        
textfile = open("features_final.txt", "w")
for element in final_feats:
    textfile.write(element + "\n")
textfile.close()

final_feats

# Import data

X_test = pd.read_excel(f'../processed_data/outliers_removed/imputed/knn/x_test.xlsx')
X_train = pd.read_excel(f'../processed_data/outliers_removed/imputed/knn/x_train.xlsx')
y_train = pd.read_excel(f'../processed_data/outliers_removed/imputed/knn/y_train.xlsx')['mean_happiness_binary'].values
y_test = pd.read_excel(f'../processed_data/outliers_removed/imputed/knn/y_test.xlsx')

# Wide ranging feature no tests

model_names = ['Logistic Regression', 'SVM ovo', 'SVM linear', 'Random Forest', 'Neural Network','Decision Tree','Gradient Boost',
 'ExtremeGradientBoost', 'K nearest neighbours','Light Gradient Boost','Cat Boost', 'Gaussian Native Bayes']

comparison_df = pd.DataFrame(index=model_names)

for i in range(3, 35, 5):
    print(i)
    X_test_filtered = X_test[final_feats[:i]]
    X_train_filtered = X_train[final_feats[:i]]

    all_models = make_models(X_train_filtered, y_train)

    models_df = compare_models(all_models, X_train_filtered, y_train, X_test_filtered, y_test)
    comparison_df[f'{i} features'] = models_df.set_index('Name')['Accuracy']
    
comparison_df

# Smaller ranging feature number tests

comparison_df = pd.DataFrame(index=model_names)

for i in range(1, 11):
    print(i)
    X_test_filtered = X_test[final_feats[:i]]
    X_train_filtered = X_train[final_feats[:i]]

    all_models = make_models(X_train_filtered, y_train)

    models_df = compare_models(all_models, X_train_filtered, y_train, X_test_filtered, y_test)
    comparison_df[f'{i} features'] = models_df.set_index('Name')['Accuracy']
    
comparison_df

6 and 9 features are good
