==> Notebooks/07_kfold_cross_validation.ipynb <==
from numpy import array
from sklearn.model_selection import KFold
from sklearn.model_selection import cross_val_score
from sklearn.impute import SimpleImputer
from sklearn.linear_model import LogisticRegression
from sklearn.ensemble import RandomForestClassifier
from sklearn.ensemble import GradientBoostingClassifier
from sklearn import svm
import pandas as pd
import statistics
import xgboost as xgb
from xgboost import XGBClassifier
from sklearn.impute import KNNImputer
from sklearn.metrics import roc_auc_score
from sklearn import metrics
import numpy as np

# data sample
data = pd.read_csv('../clustering_data_with_noout_cols.csv')

with open("features_final.txt") as f:
    features = f.readlines()
    features = [f.replace('\n', '') for f in features]
    

features = features[:5]
features

# prepare cross validation
kfold = KFold(n_splits=10, shuffle=True, random_state=42)

imputed_full = SimpleImputer(strategy='most_frequent').fit_transform(data)
imputed_full = pd.DataFrame(imputed_full, columns=data.columns)

RF_scores = []
LR_scores = []
SVM_scores = []
XGB_scores = []
GB_scores = []

RF_aucs = []
LR_aucs = []
SVM_aucs = []
XGB_aucs = []
GB_aucs = []

RF_cm = []
LR_cm = []
SVM_cm = []
XGB_cm = []
GB_cm = []

i=0

# enumerate splits
for train, test in kfold.split(data):
    y_train = data['mean_happiness_binary'].iloc[train]
    X_train = data[features].iloc[train]
    y_test = data['mean_happiness_binary'].iloc[test]
    imputer = KNNImputer(n_neighbors=50, weights='distance')

    # imputing training
    imputer.fit(X_train)
    X_train_knn = imputer.transform(X_train)
    x_train_imputed = pd.DataFrame(X_train_knn, columns=X_train.columns)

    # imputing test data
    feature_data = data[features]
    X_full_imputed = imputer.fit_transform(feature_data)
    X_full_knn_df = pd.DataFrame(X_full_imputed, columns=feature_data.columns)
    x_test_imputed = X_full_knn_df.iloc[test]

    #run model
    LR = LogisticRegression(solver='lbfgs', multi_class='multinomial')
    LR.fit(x_train_imputed, y_train)
    LR_scores.append(LR.score(x_test_imputed, y_test))
    LR_y_pred = LR.predict(x_test_imputed)
    LR_aucs.append(roc_auc_score(y_test, LR_y_pred))
    LR_cm.append(metrics.confusion_matrix(y_test, LR_y_pred))

    RF = RandomForestClassifier(n_estimators=1000, max_depth=10)
    RF.fit(x_train_imputed, y_train)
    RF_scores.append(RF.score(x_test_imputed, y_test))
    RF_y_pred = RF.predict(x_test_imputed)
    RF_aucs.append(roc_auc_score(y_test, RF_y_pred))
    RF_cm.append(metrics.confusion_matrix(y_test, RF_y_pred))


    SVM_ovo = svm.SVC(kernel="linear")
    SVM_ovo.fit(x_train_imputed, y_train)
    SVM_scores.append(SVM_ovo.score(x_test_imputed, y_test))
    SVM_y_pred = SVM_ovo.predict(x_test_imputed)
    SVM_aucs.append(roc_auc_score(y_test, SVM_y_pred))
    SVM_cm.append(metrics.confusion_matrix(y_test, SVM_y_pred))

    GB = GradientBoostingClassifier(n_estimators=1000, max_depth=10)
    GB.fit(x_train_imputed, y_train)
    GB_scores.append(GB.score(x_test_imputed, y_test))
    GB_y_pred = GB.predict(x_test_imputed)
    GB_aucs.append(roc_auc_score(y_test, GB_y_pred))
    GB_cm.append(metrics.confusion_matrix(y_test, GB_y_pred))
    
    print(i)
    i+=1

# Accuracies

print('Linear Regression')
print(statistics.mean(LR_scores))

print('\nRandom Forest')
print(statistics.mean(RF_scores))

print('\nSVM ono')
print(statistics.mean(SVM_scores))

print('\nGradient Boost')
print(statistics.mean(GB_scores))

# AUC ROC

print('Linear Regression')
print(statistics.mean(LR_aucs))

print('\nRandom Forest')
print(statistics.mean(RF_aucs))

print('\nSVM ono')
print(statistics.mean(SVM_aucs))

print('\nGradient Boost')
print(statistics.mean(GB_aucs))

# Confusion Matrix

'''cm = ConfusionMatrix(y_test.to_numpy(), y_pred, digit=5)
cm.save_html('RF_pycm2.html')

from pycm import*

from pandas_profiling import ProfileReport
profile = ProfileReport(df_num, title="Pandas Profiling Report",vars={"num":{"low_categorical_threshold":0}})
profile

'''

for cm in [RF_cm, LR_cm, SVM_cm, GB_cm]:
    cm_flat = [testcase.ravel() for testcase in cm]
    tn, fp, fn, tp = np.mean(cm_flat, axis=0)
    print(tp, tn, fp, fn)

# Exploration

from pycm import *
cm = ConfusionMatrix(y_test.to_numpy(), RF_y_pred, digit=5)
cm.save_html('RF.html')
